Fix: 사이드네비게이션 모바일 반응형 레이아웃 변경
