+++ src/pages/my/reservation-history/index.tsx
@@ -3,6 +3,7 @@ import Card from '@/components/Card';
 import MyReservationCardInfo from '@/components/Card/myReservationCardInfo';
 import FilterDropButton from '@/components/FilterButton/FilterDropButton';
 import SideNavigation from '@/components/SideNavigation';
+import SideNavigationMobileSize from '@/components/SideNavigation/MobileSize';
 import { IReservationCardInfo } from '@/types/ReservationInfo';
 import { StatusFilter } from '@/types/StatusFilter';
 import setReservationStatueInfo from '@/utils/setReservationStatueInfo';
@@ -17,6 +18,12 @@ function ReservationHistory() {
   const { ref, inView } = useInView({
     threshold: 0,
   });
+
+  const [isSideNavVisible, setIsSideNavVisible] = useState(false);
+
+  const toggleSideNav = () => {
+    setIsSideNavVisible(!isSideNavVisible);
+  };
 
   const handleGetReservationList = useCallback(async (filterOption: string | undefined) => {
     const result = await getMyReservationList({ filterOption });
@@ -53,7 +60,12 @@ function ReservationHistory() {
   return (
     <>
       <section className='pt-[14.2rem] pb-[15rem] px-[2rem] max-w-[124rem] mx-auto flex gap-[2.4rem] items-start'>
-        <SideNavigation />
+        <div className='hidden md:block'>
+          <SideNavigation />
+        </div>
+        <div className='md:hidden '>
+          <SideNavigationMobileSize toggleSideNav={toggleSideNav} isSideNavVisible={isSideNavVisible} />
+        </div>
         <div className='w-full'>
           <div className='flex justify-between items-center'>
             <h1 className='text-[3.2rem] font-[700] leading-[3.819rem]'>예약 내역</h1>
